Refactor server code to decouple from play as much as possible.
Review: @joshmutus
