+++ server/src/main/resources/routes
@@ -1,16 +1,16 @@
-GET     /api/socket    org.labrad.browser.JsonRpcController.socket
+GET     /api/socket    org.labrad.browser.BrowserController.socket
 
 # for all app URLs just serve index.html; js code renders the right UI
-GET     /                 org.labrad.browser.BrowserController.index(path="")
-GET     /dataset          org.labrad.browser.BrowserController.index(path="")
-GET     /dataset*path     org.labrad.browser.BrowserController.index(path)
-GET     /grapher          org.labrad.browser.BrowserController.index(path="")
-GET     /grapher*path     org.labrad.browser.BrowserController.index(path)
-GET     /nodes            org.labrad.browser.BrowserController.index(path="")
-GET     /nodes*path       org.labrad.browser.BrowserController.index(path)
-GET     /registry         org.labrad.browser.BrowserController.index(path="")
-GET     /registry*path    org.labrad.browser.BrowserController.index(path)
-GET     /server/*path     org.labrad.browser.BrowserController.index(path)
+GET     /                 org.labrad.browser.BrowserController.app(path="")
+GET     /dataset          org.labrad.browser.BrowserController.app(path="")
+GET     /dataset*path     org.labrad.browser.BrowserController.app(path)
+GET     /grapher          org.labrad.browser.BrowserController.app(path="")
+GET     /grapher*path     org.labrad.browser.BrowserController.app(path)
+GET     /nodes            org.labrad.browser.BrowserController.app(path="")
+GET     /nodes*path       org.labrad.browser.BrowserController.app(path)
+GET     /registry         org.labrad.browser.BrowserController.app(path="")
+GET     /registry*path    org.labrad.browser.BrowserController.app(path)
+GET     /server/*path     org.labrad.browser.BrowserController.app(path)
 
 # serve static assets
 GET     /*file            controllers.Assets.at(path="/public", file)
